Use type double in fexl_benchmark to handle negative memory usage.
The fexl_benchmark function was subtracting unsigned long values to compute the
number of steps and bytes used during evaluation.  In the case where memory is
freed, the number of bytes used can be negative.  To handle this case, I now
cast the values to type double before subtracting.

The number of steps can never be negative, unless somehow you run so long that
you exceed 2^64 steps and roll the counter over.  Nevertheless I handle steps
the same way as bytes.

+++ src/type_run.c
@@ -295,8 +295,8 @@ value type_fexl_benchmark(value f)
 
 	{
 	value x = arg(f->L->R);
-	unsigned long steps = num_steps - save_num_steps;
-	unsigned long bytes = cur_bytes - save_cur_bytes;
+	double steps = (double)num_steps - (double)save_num_steps;
+	double bytes = (double)cur_bytes - (double)save_cur_bytes;
 	f = A(A(A(hold(f->R),x),Qnum0(steps)),Qnum0(bytes));
 	}
